+++ chain/x/plasticcredit/keeper/msg_server_test.go
@@ -2,7 +2,6 @@ package keeper_test
 
 import (
 	"math/rand"
-	"testing"
 	"time"
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -10,7 +9,6 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/authz"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/stretchr/testify/suite"
 
 	"github.com/EmpowerPlastic/empowerchain/app"
 	"github.com/EmpowerPlastic/empowerchain/app/params"
@@ -1736,28 +1734,4 @@ func (s *TestSuite) TestRetireCredits() {
 			}
 		})
 	}
-}
-
-func TestTestSuite(t *testing.T) {
-	params.SetAddressPrefixes()
-	params.RegisterDenoms()
-
-	ts := &TestSuite{}
-	ts.numTestIssuers = 2
-	ts.issuerCreator = sample.AccAddress()
-	ts.sampleIssuerID = 1
-	ts.sampleIssuerAdmin = sample.AccAddress()
-	ts.noCoinsIssuerID = 2
-	ts.noCoinsIssuerAdmin = sample.AccAddress()
-	ts.sampleCreditClassAbbreviation = "EMP"
-	ts.sampleApplicantID = 1
-	ts.sampleApplicantAdmin = sample.AccAddress()
-	ts.sampleProjectID = 1
-	ts.sampleUnapprovedProjectID = 2
-	ts.sampleRejectionProjectID = 3
-	ts.sampleSuspendedProjectID = 4
-	ts.sampleCreditDenom = "EMP/123"
-	ts.creditClassCreationFee = sdk.NormalizeCoin(plasticcredit.DefaultCreditClassCreationFee)
-
-	suite.Run(t, ts)
 }
